// src/store/dashboard/type.ts
export interface IDashboardState {
  //每个商品的个数
  goodsCount: any[]
  // 每个分类的销量
  goodsSale: any[]
  // 每个分类的收藏
  goodsFavor: any[]
  //销售前10的商品数量
  goodsSaleTop: any[]
  //不同城市的销量数据
  goodsAddressSale: any[]
  //商品数据统计的数量
  goodsAmountList: any[]
}